Create swipe_bot.py
One bot to swipe them all
